Moved group detection to the Mirabuf file
