+++ engine/Assets/Scripts/Importer/Importer.cs
@@ -36,17 +36,7 @@ namespace Synthesis.Import
 	/// The Importer class connected functions with string parameters to import an Entity/Model into the Engine
 	/// NOTE: This may be moved into the Engine instead of in it's own project.
 	/// </summary>
-	public static class Importer
-	{
-		#region Profiling
-
-		public static readonly ProfilerMarker RigidgroupParsing = new ProfilerMarker("Importer.RG_Parsing");
-		public static readonly ProfilerMarker BodyCreation = new ProfilerMarker("Importer.Body_Creation");
-		public static readonly ProfilerMarker JointCreation = new ProfilerMarker("Importer.Joint_Creation");
-		public static readonly ProfilerMarker EntireImport = new ProfilerMarker("Importer.All");
-		public static readonly ProfilerMarker CollisionIgnoring = new ProfilerMarker("Importer.Collision_Ignoring");
-
-		#endregion
+	public static class Importer {
 
 		#region Importer Framework
 
@@ -71,17 +61,12 @@ namespace Synthesis.Import
 			return MirabufAssemblyImport(new MirabufLive(path));
 		}
 
-		private static List<Collider> _collidersToIgnore;
-
 		/// <summary>
 		/// Import a mirabuf assembly from a live file.
 		/// </summary>
 		/// <param name="miraLive">Live file for mirabuf data</param>
 		/// <returns>A tuple of the main gameobject that contains the imported assembly, a reference to the live file, and the simobject controlling the assembly</returns>
-		public static (GameObject MainObject, MirabufLive MiraAssembly, SimObject Sim) MirabufAssemblyImport(MirabufLive miraLive)
-		{
-			EntireImport.Begin();
-
+		public static (GameObject MainObject, MirabufLive miraLive, SimObject Sim) MirabufAssemblyImport(MirabufLive miraLive) {
 			// Uncommenting this will delete all bodies so the JSON file isn't huge
 			// DebugAssembly(miraLive.MiraAssembly);
 			// return null;
@@ -94,11 +79,7 @@ namespace Synthesis.Import
 				Logger.Log($"Hey Dev, the assembly you're importing is using a higher version than the current set version. Please update the CURRENT_MIRA_EXPORTER_VERSION constant", LogLevel.Debug);
 			}
 
-			// Logger.Log((new System.Diagnostics.StackTrace()).ToString(), LogLevel.Debug);
-
 			UnityEngine.Physics.sleepThreshold = 0;
-
-			RigidgroupParsing.Begin();
 
 			GameObject assemblyObject = new GameObject(assembly.Info.Name);
 			var parts = assembly.Data.Parts;
@@ -107,19 +88,14 @@ namespace Synthesis.Import
 			var groupObjects = new Dictionary<string, GameObject>();
 			var jointToJointMap = new Dictionary<string, (UnityEngine.Joint a, UnityEngine.Joint b)>();
 			float totalMass = 0;
-			_collidersToIgnore = new List<Collider>();
 
-			// Import Rigid Definitions
 			#region Rigid Definitions
 
 			var gamepieces = new List<GamepieceSimObject>();
-			var rigidDefinitions = FindRigidbodyDefinitions(assembly);
+			var rigidDefinitions = miraLive.Definitions;
 
-			RigidgroupParsing.End();
-			BodyCreation.Begin();
+			foreach (var group in rigidDefinitions.Definitions.Values) {
 
-			foreach (var group in rigidDefinitions.definitions.Values)
-			{
 				GameObject groupObject = new GameObject(group.Name);
 				var collectivePhysData = new List<(UnityEngine.Transform, MPhysicalProperties)>();
 				var isGamepiece = !assembly.Dynamic && group.Name.Contains("gamepiece");
@@ -137,9 +113,7 @@ namespace Synthesis.Import
 						MakePartDefinition(partObject, partDefinition, partInstance, assembly.Data, !isGamepiece, !isStatic);
 						partObjects.Add(partInstance.Info.GUID, partObject);
 						partObject.transform.parent = groupObject.transform;
-						// MARK: If transform changes do work recursively, apply transformations here instead of in a separate loop
 						var gt = partInstance.GlobalTransform.CorrectUnityMatrix;
-						// Debug.Log(gt.GetPosition(mod: true));
 						partObject.transform.localPosition = gt.GetPosition();
 						partObject.transform.localRotation = gt.rotation;
 						// partObject.transform.ApplyMatrix(partInstance.GlobalTransform);
@@ -149,12 +123,14 @@ namespace Synthesis.Import
 					}
 				}
 
+				groupObject.transform.parent = assemblyObject.transform;
+
+				#endregion
+
 				if (!assembly.Dynamic && !isGamepiece)
 					groupObject.transform.GetComponentsInChildren<UnityEngine.Transform>().ForEach(x => x.gameObject.layer = FIELD_LAYER);
 				else if (assembly.Dynamic)
 					groupObject.transform.GetComponentsInChildren<UnityEngine.Transform>().ForEach(x => x.gameObject.layer = DYNAMIC_1_LAYER);
-
-				#endregion
 
 				// Combine all physical data for grouping
 				var combPhysProps = CombinePhysicalProperties(collectivePhysData);
@@ -164,7 +140,7 @@ namespace Synthesis.Import
 				rb.mass = (float) combPhysProps.Mass;
 				totalMass += rb.mass;
 				rb.centerOfMass = combPhysProps.Com; // I actually don't need to flip this
-				groupObject.transform.parent = assemblyObject.transform;
+				
 				if (isGamepiece) {
 					var gpSim = new GamepieceSimObject(group.Name, groupObject);
 					try {
@@ -182,13 +158,9 @@ namespace Synthesis.Import
 				// DebugJointAxes.DebugPoints.Add((combPhysProps.Com, () => groupObject.transform.localToWorldMatrix));
 			}
 
-			BodyCreation.End();
-
 			#endregion
 
 			#region Joints
-
-			JointCreation.Begin();
 
 			var state = new ControllableState
 			{
@@ -230,10 +202,10 @@ namespace Synthesis.Import
 
 				// Logger.Log($"Joint Instance: {jointKvp.Key}", LogLevel.Debug);
 				// Logger.Log($"Parent: {jointKvp.Value.ParentPart}", LogLevel.Debug);
-				var aKey = rigidDefinitions.partToDefinitionMap[jointKvp.Value.ParentPart];
+				var aKey = rigidDefinitions.PartToDefinitionMap[jointKvp.Value.ParentPart];
 				var a = groupObjects[aKey];
 				// Logger.Log($"Child: {jointKvp.Value.ChildPart}", LogLevel.Debug);
-				var bKey = rigidDefinitions.partToDefinitionMap[jointKvp.Value.ChildPart];
+				var bKey = rigidDefinitions.PartToDefinitionMap[jointKvp.Value.ChildPart];
 				var b = groupObjects[bKey];
 
 				MakeJoint(
@@ -247,22 +219,6 @@ namespace Synthesis.Import
 				);
 			}
 
-			JointCreation.End();
-
-			CollisionIgnoring.Begin();
-
-			// TODO: This is really slow
-			// for (var i = 0; i < _collidersToIgnore.Count/* - 1*/; i++)
-			// {
-			// 	_collidersToIgnore[i].gameObject.layer = 7;
-			// 	// for (var j = i + 1; j < _collidersToIgnore.Count; j++)
-			// 	// {
-			// 	// 	UnityEngine.Physics.IgnoreCollision(_collidersToIgnore[i], _collidersToIgnore[j]);
-			// 	// }
-			// }
-
-			CollisionIgnoring.End();
-
 			#endregion
 
 			if (assembly.Dynamic) {
@@ -270,10 +226,10 @@ namespace Synthesis.Import
 				// (simObject as RobotSimObject).ConfigureTestSimulationBehaviours();
 			}
 
-			EntireImport.End();
-
 			return (assemblyObject, miraLive, simObject);
 		}
+
+		#endregion
 
 		#region Assistant Functions
 
@@ -293,8 +249,6 @@ namespace Synthesis.Import
 
 					var parentPartInstance = assembly.Data.Parts.PartInstances[instance.ParentPart];
 					var parentPartDefinition = assembly.Data.Parts.PartDefinitions[parentPartInstance.PartDefinitionReference];
-
-					// var moddedMat = parentPartInstance.GlobalTransform.UnityMatrix;
 
 					var originA = (UVector3)(definition.Origin ?? new UVector3());
 					// var firstPart = assembly.Data.Parts.PartInstances.First(x => x.Value.Info.Name.Equals(parentPartDefinition.Info.Name + ":1")).Value;
@@ -323,7 +277,6 @@ namespace Synthesis.Import
 					}
 
 					revoluteA.axis = axisWut;
-						// ((UVector3)definition.Rotational.RotationalFreedom.Axis).normalized;
 					revoluteA.connectedBody = rbB;
 					revoluteA.connectedMassScale = revoluteA.connectedBody.mass / rbA.mass;
 					revoluteA.useMotor = true;
@@ -331,10 +284,6 @@ namespace Synthesis.Import
 					var limits = definition.Rotational.RotationalFreedom.Limits;
 					if (limits != null && limits.Lower != limits.Upper) {
 						revoluteA.useLimits = true;
-						// revoluteA.limits = new JointLimits() {
-						// 	min = limits.Lower * Mathf.Rad2Deg,
-						// 	max = limits.Upper * Mathf.Rad2Deg
-						// };
 						revoluteA.limits = new JointLimits() {
 							min = -limits.Upper * Mathf.Rad2Deg,
 							max = -limits.Lower * Mathf.Rad2Deg
@@ -383,24 +332,6 @@ namespace Synthesis.Import
 					sliderA.angularXMotion = ConfigurableJointMotion.Locked;
 					sliderA.angularYMotion = ConfigurableJointMotion.Locked;
 					sliderA.angularZMotion = ConfigurableJointMotion.Locked;
-					// switch (definition.Prismatic.PrismaticFreedom.PivotDirection) {
-					// 	case Axis.Y:
-					// 		sliderA.xMotion = ConfigurableJointMotion.Locked;
-					// 		sliderA.yMotion = ConfigurableJointMotion.Limited;
-					// 		sliderA.zMotion = ConfigurableJointMotion.Locked;
-					// 		break;
-					// 	case Axis.Z:
-					// 		sliderA.xMotion = ConfigurableJointMotion.Locked;
-					// 		sliderA.yMotion = ConfigurableJointMotion.Locked;
-					// 		sliderA.zMotion = ConfigurableJointMotion.Limited;
-					// 		break;
-					// 	case Axis.X:
-					// 	default:
-					// 		sliderA.xMotion = ConfigurableJointMotion.Limited;
-					// 		sliderA.yMotion = ConfigurableJointMotion.Locked;
-					// 		sliderA.zMotion = ConfigurableJointMotion.Locked;
-					// 		break;
-					// }
 					sliderA.xMotion = ConfigurableJointMotion.Limited;
 					sliderA.yMotion = ConfigurableJointMotion.Locked;
 					sliderA.zMotion = ConfigurableJointMotion.Locked;
@@ -416,31 +347,11 @@ namespace Synthesis.Import
 					sliderB.angularXMotion = ConfigurableJointMotion.Locked;
 					sliderB.angularYMotion = ConfigurableJointMotion.Locked;
 					sliderB.angularZMotion = ConfigurableJointMotion.Locked;
-					// switch (definition.Prismatic.PrismaticFreedom.PivotDirection) {
-					// 	case Axis.Y:
-					// 		sliderB.xMotion = ConfigurableJointMotion.Locked;
-					// 		sliderB.yMotion = ConfigurableJointMotion.Free;
-					// 		sliderB.zMotion = ConfigurableJointMotion.Locked;
-					// 		break;
-					// 	case Axis.Z:
-					// 		sliderB.xMotion = ConfigurableJointMotion.Locked;
-					// 		sliderB.yMotion = ConfigurableJointMotion.Locked;
-					// 		sliderB.zMotion = ConfigurableJointMotion.Free;
-					// 		break;
-					// 	case Axis.X:
-					// 	default:
-					// 		sliderB.xMotion = ConfigurableJointMotion.Free;
-					// 		sliderB.yMotion = ConfigurableJointMotion.Locked;
-					// 		sliderB.zMotion = ConfigurableJointMotion.Locked;
-					// 		break;
-					// }
 					sliderB.xMotion = ConfigurableJointMotion.Free;
 					sliderB.yMotion = ConfigurableJointMotion.Locked;
 					sliderB.zMotion = ConfigurableJointMotion.Locked;
 					sliderB.connectedBody = rbA;
 					sliderB.connectedMassScale = sliderB.connectedBody.mass / rbB.mass;
-					// var ulimitB = sliderA.linearLimit;
-					// ulimitB.limit = midRange;
 
 					if (instance.HasSignal()) {
 						var slideDriver = new LinearDriver(assembly.Data.Signals.SignalMap[instance.SignalReference].Info.GUID,
@@ -528,8 +439,8 @@ namespace Synthesis.Import
 
 					if (collider != null)
 						collider.material = physMat;
-						if (addToColliderIgnore)
-							_collidersToIgnore.Add(collider);
+						// if (addToColliderIgnore)
+						// 	_collidersToIgnore.Add(collider);
 				}
 				bodyObject.transform.parent = container.transform;
 				// Ensure all transformations are zeroed after assigning parent
@@ -549,337 +460,10 @@ namespace Synthesis.Import
 			return new MPhysicalProperties { Mass = total, Com = com };
 		}
 
-		#region New RigidDefinition Gathering
+		public static Dictionary<string, Matrix4x4> MakeGlobalTransformations(Assembly assembly) {
 
-		/// <summary>
-		/// I really don't like how I made this. It gets the job done but I feel like it
-		/// could use a ton of optimizations.
-		/// TODO: Maybe get rid of the dictionary in the rigidbodyDefinitions and just
-		/// 	store keys. I feel like that would be a bit better.
-		/// </summary>
-		/// <param name="definitions"></param>
-		/// <param name="assembly"></param>
-		/// <returns></returns>
-		private static (
-			Dictionary<string, RigidbodyDefinition> definitions,
-			Dictionary<string, string> partToDefinitionMap
-			) FindRigidbodyDefinitions(Assembly assembly)
-		{
-			var defs = new Dictionary<string, RigidbodyDefinition>();
-			var partMap = new Dictionary<string, string>();
-
-			// Create grounded node
-			// var groundedJoint = assembly.Data.Joints.JointInstances["grounded"];
-
-			// I'm using lambda functions so I can reuse repeated logic while taking advantage of variables in the current scope
-			Action<RigidbodyDefinition, RigidbodyDefinition> MergeDefinitions = (keep, delete) => {
-				delete.Parts.ForEach((k, v) => keep.Parts.Add(k, v));
-				for (int i = 0; i < partMap.Keys.Count; i++) {
-					if (partMap[partMap.Keys.ElementAt(i)].Equals(delete.GUID))
-						partMap[partMap.Keys.ElementAt(i)] = keep.GUID;
-				}
-				defs.Remove(delete.GUID);
-			};
-			Action<string, string, string> MoveToDef = (part, from, to) => {
-				if (from != string.Empty)
-					defs[from].Parts.Remove(part);
-				var _a = assembly.Data.Parts.PartInstances[part];
-				defs[to].Parts.Add(part, _a);
-				partMap[part] = to;
-			};
-
-			var paths = LoadPartTreePaths(assembly.DesignHierarchy);
-
-			var discoveredRigidGroups = new List<RigidGroup>();
-
-			(string guid, JointInstance joint) groundJoint = default;
-			int counter = 0;
-
-			
-
-			// Create initial definitions
-			foreach (var jInst in assembly.Data.Joints.JointInstances) {
-				RigidbodyDefinition mainDef;
-				RigidbodyDefinition stubDef;
-				bool isGrounded = jInst.Key.Equals("grounded");
-				if (isGrounded) {
-					groundJoint = (jInst.Key, jInst.Value);
-					continue; // Skip grounded to apply it last
-				}
-
-				mainDef = new RigidbodyDefinition {
-					GUID = $"{counter}",
-					Name = $"{jInst.Value.Info.Name}_Rigidgroup{counter}",
-					Parts = new Dictionary<string, PartInstance>()
-				};
-				defs[mainDef.GUID] = mainDef;
-
-				stubDef = new RigidbodyDefinition {
-						GUID = $"{counter}-2",
-						Name = $"{jInst.Value.Info.Name}_Rigidgroup{counter}-2",
-						Parts = new Dictionary<string, PartInstance>()
-				};
-				defs[stubDef.GUID] = stubDef;
-
-				// I'm slowly turning into Nick
-				// Grab all the parts listed under the joint
-				if ((jInst.Value.Parts != null && jInst.Value.Parts.Nodes != null) && jInst.Value.Parts.Nodes.Count > 0) {
-					// I don't know why AllTreeElements sometimes returns null elements
-					var tmpParts = jInst.Value.Parts.Nodes.AllTreeElements().Where(x => x.Value != null && x.Value != string.Empty)
-						.Select(x => (x.Value, assembly.Data.Parts.PartInstances[x.Value]));
-					var tmpRG = new RigidGroup {
-						Name = $"discovered_{counter}"
-					};
-					tmpRG.Occurrences.Add(tmpParts.Select(x => x.Value));
-					discoveredRigidGroups.Add(tmpRG);
-					// tmpParts.ForEach(x => MoveToDef(x.Value, string.Empty, mainDef.GUID));
-				}
-
-				// Grab all parts eligable via design hierarchy
-				var parentSiblings = IdentifyParentSiblings(paths, jInst.Value.ParentPart, jInst.Value.ChildPart);
-				// Gonna make an action so I don't have to have the same code written twice
-				Action<string, string> inherit = (part, definition) => {
-					List<Node> toCheck = new List<Node>();
-					List<Node> tmp = new List<Node>();
-					string originalDef;
-					var path = paths[part];
-					// Logger.Log($"{part}: Path length {path.Count}");
-					Node n = assembly.DesignHierarchy.Nodes[0];
-					for (int i = 0; i < path.Count; i++) {
-						n = n.Children.Find(y => y.Value.Equals(path[i]));
-					}
-					toCheck.Add(n.Children.Find(y => y.Value.Equals(part)));
-					// if (path.Count == 0) {
-					// 	toCheck.Add(assembly.DesignHierarchy.Nodes.Find(y => y.Value == part));
-					// } else {
-
-					var defObj = defs[definition];
-					defObj.Name += $"_{assembly.Data.Parts.PartInstances[part]}";
-						
-					// }
-					if (partMap.ContainsKey(part))
-						originalDef = partMap[part];
-					else
-						originalDef = string.Empty;
-
-					while (toCheck.Count > 0) {
-						toCheck.ForEach(y => {
-							partMap.TryGetValue(y.Value, out string existingDef);
-							if (existingDef == null) {
-								var _partInst = assembly.Data.Parts.PartInstances[y.Value];
-								defs[definition].Parts.Add(y.Value, _partInst);
-								partMap[y.Value] = definition;
-								tmp.AddRange(y.Children.Where(z => z != null && z.Value != null)); // Just to make sure
-							} else if (existingDef.Equals(originalDef)) {
-								MoveToDef(y.Value, originalDef, definition);
-								tmp.AddRange(y.Children.Where(z => z != null && z.Value != null)); // Just to make sure
-							}
-						});
-						toCheck.Clear();
-						toCheck.AddRange(tmp);
-						tmp.Clear();
-					}
-				};
-				inherit(parentSiblings.child, stubDef.GUID);
-				inherit(parentSiblings.parent, mainDef.GUID);
-				counter++;
-			}
-
-			// Apply Ground Last
-			if (groundJoint == default) {
-				Logger.Log("Not sure this is possible but failed to find grounded joint", LogLevel.Error);
-				throw new Exception();
-			}
-			RigidbodyDefinition groundDef = new RigidbodyDefinition {
-				Name = "grounded",
-				GUID = "grounded",
-				Parts = new Dictionary<string, PartInstance>()
-			};
-			defs.Add(groundDef.GUID, groundDef);
-			groundJoint.joint.Parts?.Nodes?.AllTreeElements().Where(x => x.Value != null && x.Value != string.Empty).ForEach(x => {
-				if (!partMap.ContainsKey(x.Value))
-					MoveToDef(x.Value, string.Empty, groundDef.GUID);
-			});
-
-			// Add disjointed to grounded
-			assembly.Data.Parts.PartInstances.Where(x => !partMap.ContainsKey(x.Key)).ForEach(x => {
-				MoveToDef(x.Key, string.Empty, "grounded");
-			});
-
-			// Check if the original grounded object has been eaten by one of the joints
-			var swallower = partMap[assembly.Data.Joints.JointInstances["grounded"].Parts.Nodes.ElementAt(0).Value];
-			if (swallower != groundDef.GUID) {
-				var def = defs[swallower];
-				MergeDefinitions(groundDef, def);
-			}
-
-			// Apply RigidGroups
-			discoveredRigidGroups.AddRange(assembly.Data.Joints.RigidGroups);
-			discoveredRigidGroups.Where(x => x.Occurrences.Count > 1).ForEach(x => {
-				RigidbodyDefinition rigidDef = new RigidbodyDefinition {
-					Name = $"{x.Name}",
-					GUID = $"{counter}",
-					Parts = new Dictionary<string, PartInstance>()
-				};
-				defs.Add(rigidDef.GUID, rigidDef);
-				x.Occurrences.ForEach(y => {
-					partMap.TryGetValue(y, out string existingDef);
-					if (existingDef == null) {
-						MoveToDef(y, string.Empty, rigidDef.GUID);
-					} else if (!existingDef.Equals(rigidDef.GUID)) {
-						if (existingDef.Equals("grounded")) {
-							MergeDefinitions(groundDef, rigidDef);
-							rigidDef = groundDef;
-						} else {
-							// Logger.Log($"MERGING DEFINITION: {defs[existingDef].Name} based on {x.Name}", LogLevel.Debug);
-							rigidDef.Name += $"_{defs[existingDef].Name}";
-							MergeDefinitions(rigidDef, defs[existingDef]);
-						}
-					}
-				});
-
-				counter++;
-			});
-
-			bool wasRemoved = false;
-			RigidbodyDefinition newGrounded = default;
-			// Clear excess rigidbodies
-			for (int i = 0; i < defs.Keys.Count; i++) {
-				if (defs[defs.Keys.ElementAt(i)].Parts.Count == 0) {
-					if (defs.Keys.ElementAt(i).Equals("grounded")) {
-						var groundedPartGuid = assembly.Data.Joints.JointInstances.Find(x => x.Key.Equals("grounded")).Value.Parts.Nodes[0].Value;
-						newGrounded = defs[partMap[groundedPartGuid]];
-						wasRemoved = true;
-					}
-					defs.Remove(defs.Keys.ElementAt(i));
-					i -= 1;
-				}
-			}
-
-			// Re-add grounded if it was removed
-			if (wasRemoved) {
-				defs.Remove(newGrounded.GUID);
-				List<string> originalKeys = new List<string>(partMap.Keys);
-				foreach (string partKey in originalKeys) {
-					if (partMap[partKey].Equals(newGrounded.GUID))
-						partMap[partKey] = "grounded";
-				}
-				newGrounded.GUID = "grounded";
-				newGrounded.Name = "grounded";
-				defs.Add(newGrounded.GUID, newGrounded);
-			}
-
-			// Move gamepieces to separate groupings
-			int gamepieceCounter = 0;
-			if (!assembly.Dynamic) {
-				assembly.Data.Parts.PartDefinitions.Where(x => x.Value.Dynamic).ForEach(
-					x => assembly.Data.Parts.PartInstances.Where(y => y.Value.PartDefinitionReference.Equals(x.Key))
-						.ForEach(y => {
-							RigidbodyDefinition rigidDef = new RigidbodyDefinition {
-								Name = $"gamepiece_{gamepieceCounter}",
-								GUID = $"{counter}",
-								Parts = new Dictionary<string, PartInstance>()
-							};
-							defs.Add(rigidDef.GUID, rigidDef);
-							GetAllPartsInBranch(y.Key, paths, assembly.DesignHierarchy.Nodes.ElementAt(0)).ForEach(z => MoveToDef(z, partMap[z], rigidDef.GUID));
-
-							gamepieceCounter++;
-							counter++;
-						})
-				);
-			}
-
-			// Make names reasonable
-			int shift = 0;
-			for (int i = 0; i < defs.Keys.Count; i++) {
-				var def = defs[defs.Keys.ElementAt(i)];
-				if (def.Name.Equals("grounded") || def.Name.Contains("gamepiece")) {
-					shift++;
-				} else {
-					def.Name = $"node_{i - shift}";
-				}
-				defs[defs.Keys.ElementAt(i)] = def;
-			}
-
-			return (defs, partMap);
-		}
-
-		public static (string parent, string child) IdentifyParentSiblings(Dictionary<string, List<string>> paths, string parent, string child) {
-			var parentAncestors = new List<string>();
-			var childAncestors = new List<string>();
-			parentAncestors.AddRange(paths[parent]);
-			childAncestors.AddRange(paths[child]);
-			if (parentAncestors.Count == 0 || childAncestors.Count == 0) {
-				return (parentAncestors.Count == 0 ? parent : parentAncestors[0], childAncestors.Count == 0 ? child : childAncestors[0]);
-			}
-			while (parentAncestors[0].Equals(childAncestors[0])) {
-				parentAncestors.RemoveAt(0);
-				childAncestors.RemoveAt(0);
-
-				if (parentAncestors.Count == 0 || childAncestors.Count == 0) {
-					return (parentAncestors.Count == 0 ? parent : parentAncestors[0], childAncestors.Count == 0 ? child : childAncestors[0]);
-				}
-			}
-			return (parentAncestors[0], childAncestors[0]);
-		}
-
-		public static Dictionary<string, List<string>> LoadPartTreePaths(GraphContainer designHierarchy) {
-			Dictionary<string, List<string>> paths = new Dictionary<string, List<string>>();
-			foreach (Node n in designHierarchy.Nodes[0].Children) {
-				LoadPartTreePaths(n, paths);
-			}
-			return paths;
-		}
-
-		public static void LoadPartTreePaths(Node n, Dictionary<string, List<string>> paths, List<string> currentPath = null) {
-			paths[n.Value] = new List<string>();
-			if (currentPath != null && currentPath.Count > 0)
-				paths[n.Value].AddRange(currentPath);
-			// paths[n.Value] = currentPath == null ? new List<string>() : new List<string>(currentPath);
-			if (currentPath == null)
-				currentPath = new List<string>();
-			currentPath.Add(n.Value);
-			foreach (Node child in n.Children) {
-				LoadPartTreePaths(child, paths, currentPath);
-			}
-			currentPath.RemoveAt(currentPath.Count - 1);
-		}
-
-		public static List<string> GetAllPartsInBranch(string rootPart, Dictionary<string, List<string>> paths, Node rootNode) {
-			var parts = new List<string>();
-			var toCheck = new List<Node>();
-			toCheck.Add(NavigateDHPath(paths[rootPart].Append(rootPart).ToList(), rootNode));
-			while (toCheck.Count > 0) {
-				var tmp = new List<Node>();
-				toCheck.ForEach(x => {
-					parts.Add(x.Value);
-					x.Children.ForEach(y => {
-						if (y.Value != string.Empty)
-							tmp.Add(y);
-					});
-				});
-				toCheck.Clear();
-				toCheck = tmp;
-			}
-			return parts;
-		}
-
-		public static Node NavigateDHPath(List<string> path, Node rootNode) {
-			var current = rootNode;
-			while (path.Count > 0) {
-				current = current.Children.First(x => x.Value.Equals(path[0]));
-				path.RemoveAt(0);
-			}
-			return current;
-		}
-
-		#endregion
-
-		public static Dictionary<string, Matrix4x4> MakeGlobalTransformations(Assembly assembly)
-		{
 			var map = new Dictionary<string, Matrix4x4>();
-			foreach (Node n in assembly.DesignHierarchy.Nodes)
-			{
+			foreach (Node n in assembly.DesignHierarchy.Nodes) {
 				Matrix4x4 trans = assembly.Data.Parts.PartInstances[n.Value].Transform == null
 					? assembly.Data.Parts
 						.PartDefinitions[assembly.Data.Parts.PartInstances[n.Value].PartDefinitionReference]
@@ -893,8 +477,7 @@ namespace Synthesis.Import
 				MakeGlobalTransformations(map, map[n.Value], assembly.Data.Parts, n);
 			}
 
-			foreach (var kvp in map)
-			{
+			foreach (var kvp in map) {
 				assembly.Data.Parts.PartInstances[kvp.Key].GlobalTransform = map[kvp.Key];
 			}
 
@@ -917,100 +500,6 @@ namespace Synthesis.Import
 				}
 			}
 		}
-
-		#endregion
-
-		// Keeping it just incase
-		#region Old RigidDefinition Gathering
-
-		static List<Node> GatherNodes(Assembly assembly, JointInstance instance)
-		{
-			var nodes = new List<Node>(assembly.DesignHierarchy.Nodes);
-			var newRoot = new List<Node>();
-			var done = false;
-			while (nodes.Any() && !done)
-			{
-				foreach (var node in nodes)
-				{
-					if (node.Value == instance.ParentPart)
-					{
-						newRoot.Add(node);
-						done = true;
-					}
-				}
-
-				var newNodes = new List<Node>();
-				nodes.ForEach(y => newNodes.AddRange(y.Children));
-				nodes = newNodes;
-			}
-			if (!done)
-			{
-				Logger.Log($"No child parts found for node {instance.Info.Name}");
-				return new List<Node>();
-			}
-
-			var res = new List<Node>();
-			while (newRoot.Any())
-			{
-				foreach (var node in newRoot)
-				{
-					res.Add(node);
-				}
-
-				var newNodes = new List<Node>();
-				newRoot.ForEach(y => newNodes.AddRange(y.Children));
-				newRoot = newNodes;
-			}
-			return res;
-		}
-
-		public static List<(JointInstance instance, List<Node> nodes)> GatherNodes(Assembly assembly)
-		{
-			var res = assembly.Data.Joints.JointInstances.Where(p => p.Value.Info.Name != "grounded")
-				.AsParallel().Select(x => (x.Value, GatherNodes(assembly, x.Value))).ToList();
-			/*
-			var exclusionList = res.SelectMany(ji => ji.Item2.Select(x => x.Value)).ToList();
-			*/
-			var groundedJoint = assembly.Data.Joints.JointInstances.First(p => p.Value.Info.Name == "grounded");
-			var nodes = new List<Node>(assembly.DesignHierarchy.Nodes);
-			var groundedNodes = new List<Node>();
-			while (nodes.Any())
-			{
-				groundedNodes.AddRange(nodes);
-				var newNodes = new List<Node>();
-				nodes.ForEach(y => newNodes.AddRange(y.Children));
-				nodes = newNodes;
-			}
-
-			res.Insert(0, (groundedJoint.Value, groundedNodes.AsParallel().Except(res.AsParallel().SelectMany(p => p.Item2)).ToList()));
-			return res;
-		}
-
-		public static List<Node> GatherGrounded(Assembly assembly)
-		{
-			var nodes = new List<Node>(assembly.DesignHierarchy.Nodes);
-			var res = new List<Node>();
-			var exclusionList = assembly.Data.Joints.JointInstances.Where(p => p.Value.Info.Name != "grounded")
-				.Select(p => GatherNodes(assembly, p.Value).Select(n => n.Value)).SelectMany(x => x).ToList();
-			while (nodes.Any())
-			{
-				foreach (var node in nodes)
-				{
-					if (!exclusionList.Contains(node.Value))
-					{
-						res.Add(node);
-					}
-				}
-
-				var newNodes = new List<Node>();
-				nodes.ForEach(y => newNodes.AddRange(y.Children));
-				nodes = newNodes;
-			}
-
-			return res;
-		}
-
-		#endregion
 
 		#endregion
 
@@ -1067,16 +556,6 @@ namespace Synthesis.Import
 		#endregion
 
 		#region Assistant Types
-
-		/// <summary>
-		/// Collection of parts that move together
-		/// </summary>
-		public struct RigidbodyDefinition
-		{
-			public string                           GUID;
-			public string                           Name;
-			public Dictionary<string, PartInstance> Parts; // Using a dictionary to make Key searches faster
-		}
 
 		/// <summary>
 		/// Stock source type definitions
