+++ engine/Packages/manifest.json
@@ -4,7 +4,7 @@
     "com.unity.ide.rider": "3.0.15",
     "com.unity.ide.visualstudio": "2.0.16",
     "com.unity.ide.vscode": "1.2.5",
-    "com.unity.render-pipelines.universal": "10.5.1",
+    "com.unity.render-pipelines.universal": "10.8.1",
     "com.unity.test-framework": "1.1.31",
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.timeline": "1.6.4",
