+++ engine/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2020.3.12f1
-m_EditorVersionWithRevision: 2020.3.12f1 (b3b2c6512326)
+m_EditorVersion: 2020.3.28f1
+m_EditorVersionWithRevision: 2020.3.28f1 (f5400f52e03f)
